requirements-gpu.txt
vietocr
tensorflow-gpu==1.15.0
numpy
opencv-python
pillow
torch
torchvision
PyYAML
mtcnn
keras-vggface
face-alignment
mtcnn
mysql-connector
tzlocal
keras